made function available locally and renamed configuration.py to helper_functions.py
