+++ src/daemon.py
@@ -13,7 +13,7 @@ from typing import List
 from web3 import Web3
 from src.quasimodo_ebbo.on_chain_surplus import QuasimodoTestEBBO
 from src.off_chain.cow_endpoint_surplus import EndpointSolutionsEBBO
-from src.configuration import get_logger, get_tx_hashes_by_block
+from src.helper_functions import get_logger, get_tx_hashes_by_block, get_solver_competition_data
 from src.constants import INFURA_KEY
 from src.constants import SLEEP_TIME_IN_SEC
 
@@ -64,7 +64,7 @@ class DaemonEBBO:
         Function checks if solver competition data is retrievable and runs
         EBBO test, else returns True to add to list of unchecked hashes
         """
-        response_data = self.cow_endpoint_test_instance.get_solver_competition_data(
+        response_data = get_solver_competition_data(
             [single_hash]
         )
         if len(response_data) != 0:
